کالجها/جامپ بکاند با جنگو/مدرسه/requirements.txt
Django>=3.1.2
djangorestframework